feat(contact): finish contact section
Factor out education & contact sections. Change colours.
