+++ src/components/layout.css
@@ -1,4 +1,4 @@
-@import url("https://fonts.googleapis.com/css?family=Amiri&display=swap");
+@import url("https://fonts.googleapis.com/css?family=Amiri|Arimo&display=swap");
 /*
 Pantone Colours
   dark-jacaranda: #20034A;
